App now exits only when last ruler is closed
The app now closes when the last ruler is closed. It used to be that the first
ruler shown was the MainForm (I really should rename this) and thus, closing it
even when there were other rulers open would close all rulers. This wasn't the
case when closing any subsequently duplicated rulers.

Now, the app will close only when the last ruler is closed.

Signed-off-by: joce <jocelynlegault@gmail.com>