Single Instance implementation
The Ruler app will now only allow a single instance to run at any time. The
first instance that will start will act as a server for any subsequent ruler
that will start afterwards. A Ruler app that does not start first simply sends
a signal to the "main" one that will start on in its stead.

Signed-off-by: joce <jocelynlegault@gmail.com>