Length & thickness instead of width & height
We separate the actual dimensions of the ruler (length and thickness) from the
rendering values (width and height). This makes it so a ruler would still
be e.g 500x120 (L/T) no matter if it's vertical or horizontal. This will make
it easier to pass in start arguments later.

Signed-off-by: joce <jocelynlegault@gmail.com>