Add down ticks & possibility to turn ticks on/off
Add back the "down" ticks that were present in the original version. However
they are now optional, as are the "up" ticks that @sighol added to his branch.

Also, the lock icon was moved to be with the width/height/cursor labels and
this whole package moves as a block as the ruler size changes.

In addition, this whole label + icon block has been moved to the top of the
ruler for vertical rulers and will try to stay horizontal as long as the width
will allow for it. If the ruler becomes to narrow, the block will move and
become vertical.

Signed-off-by: joce <jocelynlegault@gmail.com>