Implement a better commandline parser
The previous commandline parser waht as barebone as they come. This is a poor
man's commandline parser that nontheless improve on what was already there.
While I could have gone with a 3rd party library (e.g. NDesk.Options), I didn't
want to a) bloat the code or b) require a separate dll to have to site
alongside the exe. Given that this is not a primarily commandline tool and that
this is mainly intended for users to invoke rulers with given parameters from a
batch file, I consider this is more than enough. However, if someone consider
they could improve on the implementation, I more than welcome the input.

Signed-off-by: joce <jocelynlegault@gmail.com>